+++ packages/client-core/src/components/useFullscreen.tsx
@@ -1,0 +1,7 @@
+import { createContext, useContext } from 'react'
+
+export const FullscreenContext = createContext<[boolean, (c: boolean) => void]>([false, () => {}])
+
+export const useFullscreen = () => {
+  return useContext(FullscreenContext)
+}
